Separate oauth logic from tiktok logic in manager, add tiktok post commands, update secret flags to be required together
