+++ manager/tiktok_manager.go
@@ -1,18 +1,13 @@
 package manager
 
 import (
-	"crypto/rand"
-	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"github.com/skhanal5/clip-farmer/internal/client"
-	"github.com/skhanal5/clip-farmer/internal/server"
 	"github.com/skhanal5/clip-farmer/internal/tiktok"
 	"log"
-	"math/big"
 	"os"
-	"strings"
-	"sync"
+	"time"
 )
 
 type TikTokManager struct {
@@ -23,6 +18,22 @@ func InitTikTokManager(oauthToken string) TikTokManager {
 	return TikTokManager{oauthToken: oauthToken}
 }
 
+func (t *TikTokManager) UploadVideos(directory string) {
+	dir, err := os.ReadDir(directory)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	for _, file := range dir {
+		fileInfo, err := file.Info()
+		if err != nil {
+			log.Fatal(err)
+		}
+		time.Sleep(5 * time.Second)
+		t.UploadVideo(directory + "/" + fileInfo.Name())
+	}
+}
+
 func (t *TikTokManager) UploadVideo(filepath string) {
 	file, err := os.Open(filepath)
 	if err != nil {
@@ -31,7 +42,10 @@ func (t *TikTokManager) UploadVideo(filepath string) {
 	defer file.Close()
 	stat, err := file.Stat()
 	if err != nil {
-		log.Fatal(err)
+		log.Fatalf("Tried to open file properties, received error: %s", err)
+	}
+	if stat.IsDir() {
+		log.Fatalf("Path to file: %s is not a valid file", filepath)
 	}
 	uploadVideoAsDraft(stat.Size(), file, t.oauthToken)
 }
@@ -44,20 +58,8 @@ func uploadVideoAsDraft(size int64, file *os.File, accessToken string) string {
 	return sendVideoUploadReq(file, size, response)
 }
 
-func sendVideoUploadReq(file *os.File, size int64, response tiktok.FileUploadResponse) string {
-	byteRange := fmt.Sprintf("bytes 0-%d/%d", size-1, size)
-	videoUploadReq := tiktok.BuildVideoUploadRequest(file, byteRange, response.Data.UploadURL)
-	fmt.Println(videoUploadReq)
-	res, err := client.SendRequest(videoUploadReq)
-	if err != nil {
-		panic(err)
-	}
-	return string(res)
-}
-
 func sendFileUploadReq(accessToken string, size int64) tiktok.FileUploadResponse {
 	fileUploadReq := tiktok.BuildFileUploadRequest(accessToken, size)
-	fmt.Println(fileUploadReq)
 	res, err := client.SendRequest(fileUploadReq)
 	if err != nil {
 		panic(err)
@@ -70,73 +72,12 @@ func sendFileUploadReq(accessToken string, size int64) tiktok.FileUploadResponse
 	return videoUploadRes
 }
 
-func FetchTiktokOAuth(clientKey string, clientSecret string) string {
-	codeVerifier := generateCodeVerifier(64)
-	codeChallenge := generateRawCodeChallenge(codeVerifier)
-	code := authenticateTikTokUser(clientKey, codeChallenge)
-	oauth := sendTikTokOAuthRequest(clientKey, clientSecret, code, codeVerifier)
-	writeToFile(oauth)
-	return oauth.AccessToken
-}
-
-func generateCodeVerifier(length int) string {
-	const chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-._~"
-	var result strings.Builder
-	for i := 0; i < length; i++ {
-		index, _ := rand.Int(rand.Reader, big.NewInt(int64(len(chars))))
-		result.WriteByte(chars[index.Int64()])
-	}
-	return result.String()
-}
-
-func generateRawCodeChallenge(codeVerifier string) string {
-	hash := sha256.New()
-	hash.Write([]byte(codeVerifier))
-	hashBytes := hash.Sum(nil)
-	return fmt.Sprintf("%x", hashBytes)
-}
-
-func authenticateTikTokUser(clientKey string, codeChallenge string) string {
-	authRequest := tiktok.BuildAuthenticationRequest(clientKey, codeChallenge)
-	log.Print("Invoking TikTok Login request")
-	fmt.Println(authRequest.URL.String())
-	codeChan := make(chan string)
-	serverDone := &sync.WaitGroup{}
-	serverDone.Add(1)
-	server.StartCallbackServer(serverDone, codeChan)
-	value := <-codeChan
-	return value
-}
-
-func sendTikTokOAuthRequest(clientKey string, clientSecret string, code string, codeVerifier string) tiktok.OAuthToken {
-	oauthReq := tiktok.BuildOAuthRequest(clientKey, clientSecret, code, codeVerifier)
-	log.Print("Invoking TikTok OAuth request")
-	responseBody, err := client.SendRequest(oauthReq)
+func sendVideoUploadReq(file *os.File, size int64, response tiktok.FileUploadResponse) string {
+	byteRange := fmt.Sprintf("bytes 0-%d/%d", size-1, size)
+	videoUploadReq := tiktok.BuildVideoUploadRequest(file, byteRange, response.Data.UploadURL)
+	res, err := client.SendRequest(videoUploadReq)
 	if err != nil {
 		panic(err)
 	}
-	var oauthResponse tiktok.OAuthToken
-	fmt.Println(string(responseBody))
-	err = json.Unmarshal(responseBody, &oauthResponse)
-	if err != nil {
-		panic(err)
-	}
-	return oauthResponse
-}
-
-func writeToFile(o tiktok.OAuthToken) {
-	file, err := os.Create("tiktok_oauth_resp.json")
-	if err != nil {
-		log.Print(err)
-	}
-	defer file.Close()
-
-	data, err := json.MarshalIndent(o, "", "  ")
-	if err != nil {
-		log.Print(err)
-	}
-	_, err = file.Write(data)
-	if err != nil {
-		log.Print(err)
-	}
+	return string(res)
 }
